requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
astunparse==1.6.3
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cohere==5.13.3
colorama==0.4.6
comtypes==1.4.8
cryptography==45.0.2
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.1
faiss-cpu==1.11.0
fastavro==1.9.7
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.6.0
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.2.2
groq==0.13.1
grpcio==1.68.1
gTTS==2.5.4
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.42
libclang==18.1.1
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.18.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.79.0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pillow==11.0.0
propcache==0.3.1
protobuf==5.29.1
psutil==7.0.0
pyarrow==18.1.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
pypdf==5.5.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pyttsx3==2.98
pytz==2024.2
pywin32==308
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.12.0
SQLAlchemy==2.0.41
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.6.0
tifffile==2025.1.10
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.17.2
unstructured-client==0.35.0
urllib3==2.2.3
watchdog==6.0.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
yarl==1.20.0
zstandard==0.23.0
